[FEAT] : Give each Reader 'available_cols' so that Loader can assign
Each subclass Reader now has available_cols class variable, which can be accessed from the parent class of pykrx.
The Loader can now see which data is available and when it requests multiple data, it can request to a specific Reader who has the data.

For example, pykrx has OHLCV Reader and Marketcap Reader. When the user requests close price and market cap at the same time, the Loader will assign OHLCV Reader to retrieve close price and assign Marketcap Reader for the market cap.

Combining the two results is the responsibility of the Loader.