[FEAT] : Add remove holiday by using KOSPI200 trading days (temp)
The current structure cannot remove holidays on the Reader side because any child reader not responsible for close price cannot know holidays in advance.

For a temporary solution, I used KOSPI200 index tradingdays to filter only trading days. This approach is an easy fix but using a single index/security for the standard reference of trading days is not okay.

Ultimately, I'll need some form of trading calendar but then it will require a single source of truth, which should be later discovered.